@@ -11,8 +11,8 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
-        "composer-plugin-api": "^1.0"
+        "php": "^8.0",
+        "composer-plugin-api": "^2.3"
     },
     "minimum-stability": "stable",
     "autoload": {
@@ -24,6 +24,6 @@
         "class": "aksafan\\composer\\changelog\\Plugin"
     },
     "require-dev": {
-        "composer/composer": ">=1.10.23 <2.0.0"
+        "composer/composer": "^2.5.0"
     }
 }
